refactor: move base and bind to lib
