+++ lib/base.go
@@ -1,0 +1,5 @@
+package lib
+
+type ModelBase struct {
+	ID BinaryUUID `json:"id"`
+}
