--- models/base.go
@@ -1,7 +1,0 @@
-package models
-
-import "clean-architecture/lib"
-
-type Base struct {
-	ID lib.BinaryUUID `json:"id"`
-}
